fixed gradle file
